requirements.txt
pandas
numpy
scikit-learn
streamlit
python-docx
PyPDF2
